Embrace Ruby idioms more in ApiReferenceRenderer
The code in this file didn't appear to be very consistent with modern
Ruby patterns so I've given it something of a spruce up.

I was super nervous about the risk of introducing bugs or changing
relied upon quirks so I set myself a couple of rules: try to keep the API
identical and try keep the behaviour identical.

There are a bunch of dubious things going on in this code that does make
me a bit concerned for it's long term usage. I hope in doing this it
makes this code a little more inviting for the next person to come a
long and try make it better.